requirements.txt
Package                 Version
----------------------- -------
asgiref                 3.4.1
Django                  3.2.9
django-ckeditor         6.1.0
django-js-asset         1.2.2
django-multiselectfield 0.1.12
Pillow                  8.4.0
pip                     21.3.1
pytz                    2021.3
setuptools              57.4.0
sqlparse                0.4.2
